Expose Parameters Directly & Auto-Prop assigning
Now you can expose parameters directly in the component files and you dont have to add it to props.json anymore; that will be automatically generated from a simple comment you can put in the component file. //[prop name type]

Props are also automatically assigned to variables with corresponding names in the classes, the props object is still passed down if needed.

1. Add tag above variable
2. Editor refresh (picks it up, generates new JSON file)
3. When Scene.JSON includes that parameter it can be included
4. Variable automatically assigned to classes